lang: Fix ValueOf() for structs with pointer fields
Structs that contain fields that are pointers to any other type would
cause ValueOf() to panic. reflect.Value references obtained by
value.Field() on a struct reflect.Value appear to always have CanSet()
as false, meaning ValueOf() when following the pointer would attempt to
instantiate then set the pointer to a new instance of the value that it
required, and panic because reflect disallowed setting the struct fields
through the reflect.Value references.

Work around this issue by "cloning" the reflect.Value with a new copy,
and copy the data into it before calling ValueOf(). In most cases, this
shouldn't be a significant problem because the cloning should only
happen when the reflect.Value isn't Set()able for whatever reason. In
the few cases where that path does get called, cloning the data will
increase memory usage a small amount, although it should be short-lived
and the garbage collector will make light work of it.

Add a simple string pointer test case to the ValueOf() tests to
highlight the issue and demonstrate the fix.

Signed-off-by: Joe Groocock <me@frebib.net>