test: Replace gometalinter with golangci-lint
https://github.com/alecthomas/gometalinter/issues/590

gometalinter was archived ~2 years ago and has been unmaintained since.
It hasn't compiled for a while, and is superseded by golangci-lint which
does give or take the same job.

golangci-lint is only run as a GitHub Actions task and not as part of
the main test suite (when running locally). TBD

Signed-off-by: Joe Groocock <me@frebib.net>