debug: Changes to easily reproduce deploy bug
All changes here are from my working tree and are frankly a mess. These
are not actually required to reproduce the bug, it was just the quickest
and easiest way I could think of to get a simple reproducer together.

The bug:

Start by running mgmt with a single mcl input file and letting it apply

    mgmt run lang examples/lang/docker_container0.mcl

Then make a change to the mcl file, any one of the values in the
resource struct. Re-deploy the updated file with changes.

    mgmt deploy --no-git --seeds localhost:2380 lang examples/lang/docker_container0.mcl

In my testing, most of the time I observed no CheckApply() after
deploying, which is strange because a change in the mcl should trigger a
graph change and invalidate the resource causing it to CheckApply(). It
seems a bit hit-and-miss but most of the time it would reload when I
changed one of the ports nested within the map, but it will *never*
reload when changing one of the string input variables to the
docker.parse_volume() functions. I'm wondering if that change isn't
being propagated through the function call by the engine on graph load?

Signed-off-by: Joe Groocock <me@frebib.net>